rename connection_manager, add missing file
