+++ modules/ConnectionManager.py
